rotate.xml
This code goes inside the anim folder. Create an anim folder in your res folder and then go to "File" then "New" and then click "animation resource file" make sure to name it rotate. The code for rotate is the code right here. The whole point is just to create the animation for the dice to rotate.